Add drawing utility functions
Moves drawing helper functions to a separate module so we can reuse
them to draw scenario modifications.

Had to make these functions accept `map` as the first argument to avoid
circular dependency issues.
