+++ src/mmw/js/src/core/views.js
@@ -6,6 +6,7 @@ var $ = require('jquery'),
     router = require('../router.js').router,
     Marionette = require('../../shim/backbone.marionette'),
     TransitionRegion = require('../../shim/marionette.transition-region'),
+    drawUtils = require('../draw/utils'),
     headerTmpl = require('./templates/header.html');
 
 /**
